requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
dill==0.3.8
docker-pycreds==0.4.0
executing==2.1.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.27.1
ibm-cloud-sdk-core==3.22.1
ibm-platform-services==0.59.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
iqm-client==20.13
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
# Editable install with no version control (karateclub==1.3.4)
-e c:\users\valte\onedrive - university of helsinki\desktop\karateclub-v_10304\karateclub-v_10304
kiwisolver==1.4.8
Levenshtein==0.26.1
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pbr==6.1.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.1
PyGSP==0.5.1
PyJWT==2.10.1
pylatexenc==2.10
pyparsing==3.2.1
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.1
python-louvain==0.16
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
qiskit==1.1.2
qiskit-aer==0.14.2
qiskit-iqm==15.6
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
requests_ntlm==1.3.0
rustworkx==0.15.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
sentry-sdk==2.19.2
setproctitle==1.3.4
setuptools==75.8.0
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sspilib==0.2.0
stack-data==0.6.3
stevedore==5.4.0
symengine==0.13.0
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.7.0.dev20250116+cu126
torchaudio==2.6.0.dev20250117+cu126
torchvision==0.22.0.dev20250117+cu126
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wandb==0.19.2
wcwidth==0.2.13
websocket-client==1.8.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
